pyproject.toml
[project]
name = "space-warp"
version = "0.1.0"
description = "A revolutionary multi-display window and app layout manager for macOS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.10.0",
    "pyobjc-framework-cocoa>=12.0",
    "pyobjc-framework-applicationservices>=12.0",
    "pyobjc-framework-corelocation>=12.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
space-warp = "space_warp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/space_warp"]

[dependency-groups]
dev = [
    "pyinstaller",
    "radon",
    "ruff",
    "xenon",
]
